Refactor audit tests into their own file
